modify contrast level and add comment

+++ app/src/main/java/com/eggsy/photoshop/EPhotoShop.java
@@ -8,6 +8,12 @@ import android.graphics.Bitmap;
 
 public class EPhotoShop {
 
+    /**
+     * 灰化
+     *
+     * @param inputBitmap
+     * @return
+     */
     public static Bitmap doGray(Bitmap inputBitmap) {
         Bitmap outputImage = inputBitmap;
 
@@ -25,6 +31,12 @@ public class EPhotoShop {
         return outputImage;
     }
 
+    /**
+     * 浮雕
+     *
+     * @param inputBitmap
+     * @return
+     */
     public static Bitmap doEmboss(Bitmap inputBitmap) {
         Bitmap outputImage = inputBitmap;
 
@@ -42,6 +54,12 @@ public class EPhotoShop {
         return outputImage;
     }
 
+    /**
+     * 黑白
+     *
+     * @param inputBitmap
+     * @return
+     */
     public static Bitmap doBlackWhite(Bitmap inputBitmap) {
         Bitmap outputImage = inputBitmap;
 
@@ -59,6 +77,12 @@ public class EPhotoShop {
         return outputImage;
     }
 
+    /**
+     * 底片
+     *
+     * @param inputBitmap
+     * @return
+     */
     public static Bitmap doPhotographicPlate(Bitmap inputBitmap) {
         Bitmap outputImage = inputBitmap;
 
@@ -76,6 +100,13 @@ public class EPhotoShop {
         return outputImage;
     }
 
+    /**
+     * 快速灰化
+     *
+     * @param inputBitmap
+     * @param radius
+     * @return
+     */
     public static Bitmap doFastBlur(Bitmap inputBitmap, int radius) {
         Bitmap outputImage = inputBitmap;
 
@@ -93,6 +124,12 @@ public class EPhotoShop {
         return outputImage;
     }
 
+    /**
+     * 怀旧
+     *
+     * @param inputBitmap
+     * @return
+     */
     public static Bitmap doNostalgic(Bitmap inputBitmap) {
         Bitmap outputImage = inputBitmap;
 
@@ -110,7 +147,14 @@ public class EPhotoShop {
         return outputImage;
     }
 
-    public static Bitmap doContrast(Bitmap inputBitmap, int contrast) {
+    /**
+     * 对比度
+     *
+     * @param inputBitmap
+     * @param contrast
+     * @return
+     */
+    public static Bitmap doContrast(Bitmap inputBitmap, float contrast) {
         Bitmap outputImage = inputBitmap;
 
         int width = inputBitmap.getWidth();
@@ -127,6 +171,13 @@ public class EPhotoShop {
         return outputImage;
     }
 
+    /**
+     * 亮度
+     *
+     * @param inputBitmap
+     * @param brightness
+     * @return
+     */
     public static Bitmap doBrightness(Bitmap inputBitmap, int brightness) {
         Bitmap outputImage = inputBitmap;
 
@@ -144,6 +195,13 @@ public class EPhotoShop {
         return outputImage;
     }
 
+    /**
+     * 饱和度
+     *
+     * @param inputBitmap
+     * @param level
+     * @return
+     */
     public static Bitmap doSaturation(Bitmap inputBitmap, float level) {
         Bitmap outputImage = inputBitmap;
